Split the ResultsTable into 2 different components
